refactor: Organise tests into corresponding packages
