+++ app/src/test/java/org/fossasia/openevent/app/presenter/EventsPresenterTest.java
@@ -1,4 +1,4 @@
-package org.fossasia.openevent.app;
+package org.fossasia.openevent.app.presenter;
 
 import org.fossasia.openevent.app.data.contract.IEventDataRepository;
 import org.fossasia.openevent.app.data.contract.IUtilModel;
